tomorrow

+++ src/controllers/subcategory-controller.js
@@ -46,6 +46,9 @@ export const update_subcategory = AsyncErrorHandler(async (req, res, next) => {
     if (!subcategory) {
         return CallNext("subcategory not found", 404);
     }
+    if (subcategory._id.toString() !== req?.user?._id.toString()) {
+        return CallNext("you are not authorized", 403);
+    }
     if (req.file) {
         const { secure_url, public_id } = await cloud.uploader.upload(
             req?.file?.path,
@@ -83,6 +86,9 @@ export const delete_subcategory = AsyncErrorHandler(async (req, res, next) => {
     const { id } = req.params;
     const subcategory = await subcategory_model.findById(id);
     if (!subcategory) return CallNext("subcategory not found", 404);
+    if (subcategory._id.toString() !== req?.user?._id.toString()) {
+        return CallNext("you are not authorized", 403);
+    }
     await cloud.uploader.destroy(subcategory.image["public-id"]);
     await subcategory.deleteOne();
     return res.json({ done: true, payload: "deleted successfully" });
